# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Deckwatch API
  description: |
    Kubernetes control-plane and OCI registry surface exposed by Deckwatch.

    - `/api/*` — application, deployment, namespace, and cluster management
    - `/v2/*`  — OCI Distribution Spec v1.1 (only when the embedded registry is enabled)
    - `/metrics` — Prometheus exposition
  version: "0.1.0"
  contact:
    name: Deckwatch
    url: https://github.com/kore/deckwatch
servers:
  - url: /
    description: Same-origin (served by deckwatch)
...
paths:
...
  /api/git/branches:
    get:
      tags: [Git]
      summary: List branches on a remote git repo (auth via configured secret)
      parameters:
        - in: query
          name: repo_url
          required: true
          schema: { type: string }
        - in: query
          name: token_secret
          schema: { type: string }
      responses: { "200": { description: OK } }

  # -------------------------------------------------------------- events
...